minor service restructuring
