+++ api/rpc/service/service_test.go
@@ -1,0 +1,1 @@
+package service
