+++ api/rpc/stat/stat.go
@@ -1,0 +1,1 @@
+package stat
